move scaledparameter to its own module
