+++ qcodes/parameters/__init__.py
@@ -74,7 +74,7 @@ from .group_parameter import Group, GroupParameter
 from .grouped_parameter import DelegateGroup, DelegateGroupParameter, GroupedParameter
 from .multi_channel_instrument_parameter import MultiChannelInstrumentParameter
 from .multi_parameter import MultiParameter
-from .parameter import DelegateParameter, ManualParameter, Parameter, ScaledParameter
+from .parameter import DelegateParameter, ManualParameter, Parameter
 from .parameter_base import (
     ParamDataType,
     ParameterBase,
@@ -82,6 +82,7 @@ from .parameter_base import (
     invert_val_mapping,
 )
 from .parameter_with_setpoints import ParameterWithSetpoints, expand_setpoints_helper
+from .scaled_paramter import ScaledParameter
 from .specialized_parameters import ElapsedTimeParameter, InstrumentRefParameter
 from .sweep_values import SweepFixedValues, SweepValues
